wilder/requirements.txt
pytest~=6.2.1
setuptools~=40.8.0
click~=7.1.2
requests~=2.25.1
Flask~=1.1.2
PyInquirer~=1.0.3
python-vlc~=3.0.11115
